Split pgm module into submodules
